Restructure model relationships
Have beaten my head for far too long on lazily loaded backrefs
for the  album->genre, album->prediction mappings. Essentially, the
multiple join paths have to be constructed explicitly in a join
statement. Able to add backref attribute to the albums and genres, but
the relationship is not "lazy" loaded meaning that the list of children
is returned instead of a query object

references:
https://docs.sqlalchemy.org/en/13/orm/join_conditions.html
https://docs.sqlalchemy.org/en/13/orm/relationship_api.html#sqlalchemy.orm.relationship